+++ src/public/stdstring.h
@@ -83,6 +83,6 @@ inline ISaveRestoreOps *GetStdStringDataOps()
 //-------------------------------------
 
 #define DEFINE_STDSTRING(name) \
-	{ FIELD_CUSTOM, #name, { offsetof(classNameTypedef,name), 0 }, 1, FTYPEDESC_SAVE, NULL, GetStdStringDataOps(), NULL }
+	{ FIELD_CUSTOM, #name, { static_cast<int>( offsetof(classNameTypedef,name) ), 0 }, 1, FTYPEDESC_SAVE, NULL, GetStdStringDataOps(), NULL }
 
 #endif // STDSTRING_H
